Move shared components up to admin directory
