+++ components/admin/FileUploadProgressBar/FileUploadProgressBar.scss
